Renaming of lisp files
For melpa, it is important that the package name (emacs-js) is also the
provided feature. As a result, the main lisp file must be named
emacs-js.el. The convenience code automatically installs packages but
bypasses the dependency mechanism of the package manager: as a result I
believe it is easier to move this code to a dedicated
file (emacs-js-setup.el) which is not packaged.

For consistency (and for the nameless package), it is important that the
provided feature, filename and lisp name prefixes (e.g., functions and
variables) are all the same.